updates
updated code for both old NN code and new to check error, but something is different with them a lot. Both have some weird stuff going on, with the new code having error that hops all over the place and the old code having error that will randomly get stuck with high error.