i literally don't know what is happening
the array is actually just reshaping itself. I feed it in as a normal tensor. there is literally nothing else that has information for reshaping. it reshapes in a line of code that doesnt exist. it doesn't make any sense. i actually don't even understand what is happening. the array literally spontaniously reshapes itself. even in the debugger, it just decides to reshape itself. i can't deal with this.