backend work
redid initializers so it's actually correct with decent syntax. Gave up on half-assed attempt for activators bc it's better to fix the tensor class first. The tensor class is super messy and confusing to work with so I'm just going to restart it for now to work on some important elements  to not break the original tensor class.