optim work
optim """works""" (doesn't error) but the algorithms are kind of messed up. I know the algorithm works mathematically, because it worked on the old optim build with the same method code, so it's probably some weird function call that I haven't fixed. Should be easy-ish to fix and will do soon.