added ability for multiple paths to be traced
nabla now can trace back multiple paths. However, since this usually isn't something that needs to happen and changes the time complexity of nabla, this is toggleable, and by default is off.

I lowkey didn't implement this properly, but ran out of time to fix it and the original method still works so i'll fix it later